Consolidated kube-up-docker-compose.sh and kube-up.sh into a single script. Added some checks and user output to make scripts more user friendly.

+++ kube-down.sh
@@ -1,10 +1,15 @@
 #!/bin/bash
 echo "Stopping replication controllers, services and pods..."
 kubectl stop replicationcontrollers,services,pods --all
+if [ $? != 0 ]; then
+    echo "Kubernetes already down?"
+fi
 
 cd kubernetes
-docker-compose stop
-docker-compose rm -f -v
+if [ ! -z "$(docker-compose ps -q)" ]; then
+    docker-compose stop
+    docker-compose rm -f -v
+fi
 
 k8s_containers=`docker ps -a -f "name=k8s_" -q`
